AddedIdentity,Gamezone,ProjectionTesting

+++ GoalballAnalysisSystem.API/Contracts/V1/Requests/ProjectionRequest.cs
@@ -12,7 +12,7 @@ namespace GoalballAnalysisSystem.API.Contracts.V1.Requests
         public int X2 { get; set; }
         public int Y2 { get; set; }
         public double Speed { get; set; }
-        public int GameId { get; set; }
-        public int? GamePlayerId { get; set; }
+        public long GameId { get; set; }
+        public long? GamePlayerId { get; set; }
     }
 }
